requirements.txt
torch>=2.2.0
transformers>=4.41.0
datasets>=2.19.0
Pillow>=9.4.0
tqdm>=4.66.0
hnswlib>=0.8.0
matplotlib>=3.8.0
numpy>=1.26.0
